requirements.txt
# Matrix Bot Requirements
# Python 3.10+

# Core framework
aiogram>=3.4.0

# Database support
aiosqlite>=0.19.0

# Environment variables
python-dotenv>=1.0.0

# Optional but recommended
typing-extensions>=4.9.0